requirements.txt
langchain-openai
langchain-core
langchain-community
langchain-experimental
python-dotenv
langsmith